MovieList: Correctly display movie state with empty cutlist
A movie with a zero-sized cutlist would always appear als "half seen".
This change fixes that, if there's no cut point, it means the movie
wasn't seen yet.

The problem typically occured when trimming a movie so that the program
markers were gone, and then doing a "reset playback position". This results
in a zero-sized cuts file, which is not invalid but wasn't handled
correctly.