pmt: always restart m_PAT/m_PMT when frontend state changes to ok
Even if we have already allocated a demux, it could be that
m_PAT / m_PMT have timed out (or returned invalid data) on
the previous attempt.
When that happened, we would not restart them when the
frontend lost and re-gained it's lock, because we already
allocated a demux.

This situation might occur with a rotor setup for instance,
where the frontend gets a lock before the positioner has
reached the correct orbital position.
(as might happen when there is no current measurement, or
current measurement is not reliable)
As a result, no channel would be started, and no retries
would be attempted.