Navigation: Continue playing old service when incorrect PIN
When incorrect PIN is entered for the parental check previously
the screen went black. Now the current played service continued to
play and when that service is available in the channel list it will
also keep/set the focus to it. Why should we give a black screen here
This is also valid when the parental control is in timeout.